requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
